requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
apex==0.9.10.dev0
appdirs==1.4.4
astroid==2.11.2
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
black==21.4b2
blis==0.7.5
boto3==1.20.20
botocore==1.23.20
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
config==0.5.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.24
decorator==5.1.0
dill==0.3.4
einops==0.3.2
executing==0.8.2
faiss-gpu==1.7.1.post3
filelock==3.3.1
flake8==4.0.1
flint==0.0.1
fonttools==4.28.2
fsspec==2021.10.0
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20211023
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.6.0
huggingface-hub==0.2.1
hydra-core==1.1.1
icecream==2.1.1
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
install==1.3.5
iopath==0.1.9
ipython==7.28.0
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
kornia==0.5.11
langcodes==3.3.0
lazy-object-proxy==1.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==5.2.0
murmurhash==1.0.6
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.60
packaging==21.0
pandas==1.3.5
parso==0.8.2
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.5.1
portalocker==2.3.2
preshed==3.0.6
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycodestyle==2.8.0
pydantic==1.8.2
pyDeprecate==0.3.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.13.5
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pytorch-ranger==0.1.1
pytz==2021.3
PyYAML==5.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
sentencepiece==0.1.96
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
torch==1.9.0+cu111
torch-optimizer==0.1.0
torch-tb-profiler==0.3.1
torchaudio==0.9.0
torchmetrics==0.5.1
torchvision==0.10.0+cu111
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.14.0
yacs==0.1.8
yarl==1.7.0
taming-transformers==0.0.1
lpips==0.1.4
tokenizers==0.12.1
# tokenizers==0.10.3
transformers==4.19.2
cryptography==3.4.6
setuptools==59.5.0